+++ public/app/features/trails/auto-query/getAutoQueriesForMetric.test.ts
