+++ public/app/features/trails/auto-query/getAutoQueriesForMetric.ts
